+++ R/shiny.R
@@ -162,7 +162,7 @@ renderDataTable = function(
       uiFunc = dataTableOutput,
       renderFunc = function(shinysession, name, ...) {
         domain = tempVarsPromiseDomain(outputInfoEnv, outputName = name, session = shinysession)
-
+        removeTimestampFromSnapshot(name)
         promises::with_promise_domain(domain, renderFunc())
       },
       cacheHint = cacheHint
@@ -172,7 +172,7 @@ renderDataTable = function(
       uiFunc = dataTableOutput,
       renderFunc = function(shinysession, name, ...) {
         domain = tempVarsPromiseDomain(outputInfoEnv, outputName = name, session = shinysession)
-
+        removeTimestampFromSnapshot(name)
         promises::with_promise_domain(domain, renderFunc())
       }
     )
@@ -207,6 +207,13 @@ setAll = function(lst, env) {
     assign(name, val, env)
   })
   invisible()
+}
+
+removeTimestampFromSnapshot = function(name) {
+  shiny::snapshotPreprocessInput(paste0(name, "_state"), function(value) {
+    value$time <- NULL
+    value
+  })
 }
 
 # This promise domain is needed to set/unset temporary variables in
@@ -334,10 +341,11 @@ selectCells = function(proxy, selected, ignore.selectable = FALSE) {
 addRow = function(proxy, data, resetPaging = TRUE) {
   if ((is.matrix(data) || is.data.frame(data)) && nrow(data) != 1)
     stop("'data' must be of only one row")
+  rn <- rownames(data); if (!is.null(rn)) rn <- I(rn)
   # must apply unname() after as.list() because a data.table object
   # can't be really unnamed. The names() attributes will be
   # preserved but with empty strings (see #760).
-  invokeRemote(proxy, 'addRow', list(unname(as.list(data)), I(rownames(data)), resetPaging))
+  invokeRemote(proxy, 'addRow', list(unname(as.list(data)), rn, resetPaging))
 }
 
 #' @rdname proxy
@@ -573,14 +581,15 @@ dataTableAjax = function(session, data, rownames, filter = dataTablesFilter, out
 
 sessionDataURL = function(session, data, id, filter) {
 
-  URLdecode = shinyFun('URLdecode')
   toJSON = shinyFun('toJSON')
   httpResponse = shinyFun('httpResponse')
 
   filterFun = function(data, req) {
     # DataTables requests were sent via POST
-    params = URLdecode(rawToChar(req$rook.input$read()))
+    params = rawToChar(req$rook.input$read())
+    # I don't think the browser would send out nonASCII strings, but keep it as it is
     Encoding(params) = 'UTF-8'
+    # shiny::parseQueryString() calls httpuv::decodeURIComponent() internally and will handle encoding correctly
     params = shiny::parseQueryString(params, nested = TRUE)
 
     res = tryCatch(filter(data, params), error = function(e) {